requirements.txt
# Core Dependencies
PySide6>=6.6.0
# python-mpv>=1.0.4  # Removed: Using AVPlayer on macOS

# AI & Machine Learning
transformers>=4.35.0
sentencepiece>=0.1.99
sacremoses>=0.0.53
accelerate>=0.24.0
torch>=2.5.1
torchaudio>=2.5.1
silero-vad>=4.0.0
faster-whisper>=1.0.0
# llama-cpp-python - removed (not compatible with macOS ARM64)
huggingface_hub>=0.19.0

# Apple Silicon MLX (CoreML/Neural Engine acceleration)
mlx==0.20.0  # 固定版本避免 pad() API 不兼容
mlx-whisper>=0.1.0
mlx-lm==0.19.3  # 與 mlx 0.20.0 兼容

# Video & Audio Processing
ffmpeg-python>=0.2.0
pydub>=0.25.1
librosa>=0.10.0
soundfile>=0.12.1
opencv-python>=4.8.0  # Scene detection

# macOS Native (for fast thumbnail extraction on Apple Silicon)
pyobjc-framework-AVFoundation>=10.0  # Native video processing
pyobjc-framework-Quartz>=10.0        # Image saving
pyobjc-framework-AVKit>=10.0         # Native video playback (AVPlayer)
pyobjc-framework-Cocoa>=10.0         # AppKit/Foundation

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Subtitle Processing
pysrt>=1.1.2
ass>=0.5.0

# Database
SQLAlchemy>=2.0.0

# Utilities
python-dotenv>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
cryptography>=41.0.0
packaging>=21.0

# Development Tools (Optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Packaging
pyinstaller>=6.0.0

# macOS Apple Silicon 注意事項:
# - PyTorch 需使用 CPU 版本 (MPS 內建支援)
# - 安裝: pip install torch torchaudio --index-url https://download.pytorch.org/whl/cpu
# - 需安裝 Homebrew 的 mpv 和 ffmpeg: brew install mpv ffmpeg
